implement authentication route

+++ controllers/postController.js
@@ -3,7 +3,9 @@ const validationHandler = require('../validations/validationHandler');
 
 exports.index = async (req, res) => {
     try {
-        const posts = await Post.find().sort({ createdAt: -1 });
+        const posts = await Post.find()
+            .populate("user")
+            .sort({ createdAt: -1 });
         res.send(posts);
     }catch(err){
         next(err);
@@ -14,7 +16,8 @@ exports.show = async (req, res) => {
     try {
         const post = await Post.findOne({
             _id: req.params.id
-        });
+        }).populate("user");
+
         res.send(post);
     }catch(err){
         next(err);
@@ -28,6 +31,7 @@ exports.store = async (req, res, next) => {
         let post = new Post();
         post.description = req.body.description;
         post.image = req.file.filename;
+        post.user = req.user;
         post = await post.save();
         res.send(post);
     } catch (err) {
@@ -40,6 +44,12 @@ exports.update = async (req, res, next) => {
         validationHandler(req);
 
         let post = await Post.findById(req.params.id);
+        // Check if auth user is author
+        if (!post || post.user != req.user.id) {
+            const error = new Error("Wrong request");
+            error.statusCode = 400;
+            throw error;
+        }
         post.description = req.body.description;
         post = await post.save();
         res.send(post);
@@ -52,6 +62,11 @@ exports.delete = async (req, res, next) => {
     try {
 
         let post = await Post.findById(req.params.id);
+        if (!post || post.user != req.user.id) {
+            const error = new Error("Wrong request");
+            error.statusCode = 400;
+            throw error;
+        }
         await post.delete();
         res.send({ message: "success" });
     } catch (err) {
